requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.8.0
bitsandbytes>=0.42.0
accelerate>=0.25.0
datasets>=2.16.0
alpaca-py>=0.20.0
websockets>=12.0
loguru>=0.7.0
backtrader>=1.9.78
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
mcp>=0.1.0
